refactor: relocate dedicated components of `Contact Me`
