+++ src/features/contact_me/ContactMe.tsx
@@ -1,4 +1,4 @@
-import MessageMe from "@/features/contact_me/MessageMe";
+import MessageMe from "@/features/contact_me/components/MessageMe";
 import { SVGIcon } from "@/lib/helpers";
 import React from "react";
 import {
